cmds: Avoid msg deletion when station is not known
User messages were deleted when asked for a song id from a station that
was not supported. For example, in case a user would type

!raidwombotscreditcard

Wombot Wouldn't leave the message as it was in the message log, but it
would delete it without giving any reason to do so.
We can apply two different approaches:
    ·Ignore the message and leave it as it is.
    ·Delete the message and send a wombot message telling how sorry is
     is for not knowing about that station
This commit applies approach 1.

Signed-off-by: kktzi <who@chuntongo.crew>